# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Dashboard verilerini getirir",
                "operationId": "2c29ca3389addbd18db1fff0f925cfd3",
                "responses": {
                    "200": {
                        "description": "Dashboard verileri başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin Dashboard verileri getirildi."
                                        },
                                        "data": {
                                            "properties": {
                                                "toplam_kullanici": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "toplam_siparis": {
                                                    "type": "integer",
                                                    "example": 128
                                                },
                                                "toplam_gelir": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15432.9
                                                },
                                                "bugunku_siparis_sayisi": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "en_cok_satan_urunler": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "total_quantity": {
                                                                "type": "integer",
                                                                "example": 12
                                                            },
                                                            "product": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 3
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "iPhone 15 Pro"
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "example": 49999
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token formatı: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}